+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Moved map credits to map column so it don't get hidden by chart panel.
 - Fix support for `initUrls` in `startData.initSources`
 - Propagate `knownContainerUniqueIds` across references and their target.
+- Show scrollbar for story content in Safari iOS.
 - [The next improvement]
 
 #### 8.2.13 - 2022-09-01